requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
scikit-learn>=1.5.0
stable-baselines3>=2.3.0
gymnasium>=0.29.1
numpy>=1.26.0
pandas>=2.2.0
geopy>=2.4.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
googlemaps>=4.10.0
